requirements.txt
django
django-crispy-forms
django-sekizai
djangorestframework
markdown
django-filter
django-cors-headers